Commit-5-Adding Weather App and Web Server

+++ task-manager/src/routers/task.js
@@ -17,10 +17,10 @@ const auth = require("../middleware/auth.js")
 //     })
 // })
 
-router.post('/tasks', auth ,async (req, res) => {
+router.post('/tasks', auth, async (req, res) => {
     const task = new Task({
         ...req.body,
-        owner:req.user._id
+        owner: req.user._id
     })
     try {
         await task.save()
@@ -55,10 +55,35 @@ router.post('/tasks', auth ,async (req, res) => {
 // })
 
 //Refactoring GET /tasks end point
-router.get('/tasks', async (req, res) => {
+// GET /tasks?completed=true
+// GET /tasks?limit=2&skip=2
+// GET /tasks?sortBy=createdAt:desc
+router.get('/tasks', auth, async (req, res) => {
     try {
-        const tasks = await Task.find({})
-        res.send(tasks)
+        // const tasks = await Task.find({})
+        // const tasks = await Task.find({ owner: req.user._id })
+        // res.send(tasks)
+        //OR
+        //await req.user.populate('tasks').execPopulate()
+        const match ={}
+        const sort = {}
+        if(req.query.completed){
+            match.completed = req.query.completed === 'true'
+        }
+        if(req.query.sortBy){
+            parts = req.query.sortBy.split(':')
+            sort[parts[0]] = parts[1] === 'desc'? -1 : 1
+        }
+        await req.user.populate({
+            path:'tasks',
+            match,
+            options:{
+                limit:parseInt(req.query.limit),
+                skip:parseInt(req.query.skip),
+                sort
+            }
+        }).execPopulate()
+        res.send(req.user.tasks)        
     } catch (e) {
         res.status(500).send(e)
     }
@@ -78,10 +103,11 @@ router.get('/tasks', async (req, res) => {
 // })
 
 //Refactoring GET /tasks/:id end point
-router.get('/tasks/:id', async (req, res) => {
+router.get('/tasks/:id', auth, async (req, res) => {
     const _id = req.params.id
     try {
-        const task = await Task.findById(_id)
+        //const task = await Task.findById(_id)
+        const task = await Task.findOne({_id,owner:req.user._id})
         if (!task) {
             return res.status(404).send()
         }
@@ -91,31 +117,33 @@ router.get('/tasks/:id', async (req, res) => {
     }
 })
 
-router.patch('/tasks/:id',async (req,res)=>{
+router.patch('/tasks/:id',auth, async (req, res) => {
     const updates = Object.keys(req.body)
-    const allowedUpdates =['description','completed']
-    const isValidOperation = updates.every((update)=>allowedUpdates.includes(update))
-    if(!isValidOperation){
-        return res.status(400).send({error:'Invalid updates!'})
+    const allowedUpdates = ['description', 'completed']
+    const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
+    if (!isValidOperation) {
+        return res.status(400).send({ error: 'Invalid updates!' })
     }
     try {
-       //const task =  await Task.findByIdAndUpdate(req.params.id,req.body,{new:true, runValidators:true})
-       const task = await Task.findById(req.params.id)
-       updates.forEach((update)=>task[update] = req.body[update])
-       await task.save()
-       if(!task){
+        //const task =  await Task.findByIdAndUpdate(req.params.id,req.body,{new:true, runValidators:true})
+        //const task = await Task.findById(req.params.id)
+        const task = await Task.findOne({ _id: req.params.id, owner: req.user._id })
+        if (!task) {
             return res.status(404).send()
-       }
-       res.send(task)
+        }
+        updates.forEach((update) => task[update] = req.body[update])
+        await task.save()
+        res.send(task)
     } catch (e) {
         res.status(400).send(e)
     }
 })
 
-router.delete('/tasks/:id',async (req,res)=>{
+router.delete('/tasks/:id', auth ,async (req, res) => {
     const _id = req.params.id
     try {
-        const task = await Task.findByIdAndDelete(_id)
+        //const task = await Task.findByIdAndDelete(_id)
+        const task = await Task.findByIdAndDelete({_id:req.params.id,owner:req.user._id}) 
         if (!task) {
             return res.status(404).send()
         }
